+++ src/eca/llm_providers/azure.clj
@@ -30,6 +30,6 @@
                                                                  {}
                                                                  (string/split input #","))
                                                         :key api-key}}}))
-  (swap! db* assoc-in [:auth provider] nil)
+  (swap! db* dissoc :auth provider)
   (send-msg! (format "API key, url and models saved to %s" (.getCanonicalPath (config/global-config-file))))
   (f.login/login-done! ctx))